Update tests to be in right folder
