+++ services/indexing/app/pipeline.py
@@ -1,6 +1,6 @@
 # --- Imports and setup for the indexing pipeline ---
 
-from common.events import EventEnvelope
+from common.events import EventEnvelope, publish_event, new_event
 try:
     from aio_pika.abc import AbstractIncomingMessage
 except ImportError:
@@ -12,7 +12,6 @@ import aiofiles
 from sentence_transformers import SentenceTransformer
 import re
 import chromadb
-from common.events import publish_event, new_event
 import datetime
 import json
 import tiktoken
