+++ tests/integration/indexing/test_endpoints.py
@@ -1,0 +1,73 @@
+import pytest
+from pathlib import Path
+from fastapi.testclient import TestClient
+
+# Imports correctos para el servicio indexing
+from services.indexing.app.main import app
+from services.indexing.common.config import settings
+
+"""
+Two DeprecationWarnings appear during tests because FastAPI's 'on_event'
+is deprecated. They do not affect functionality and can be safely ignored.
+"""
+
+@pytest.fixture(scope="module", autouse=True)
+def setup_test_text_file():
+    """
+    Creates a test text file in /data/text before tests,
+    and deletes it after all tests.
+    """
+    data_root = Path(settings.data_root) / "text"
+    data_root.mkdir(parents=True, exist_ok=True)
+
+    test_file = data_root / "test_doc_endpoint.txt"
+    test_file.write_text(
+        "This is a document used for endpoint testing.",
+        encoding="utf-8"
+    )
+
+    print(f"[TEST] Creating test file at: {test_file}")
+    yield
+
+    if test_file.exists():
+        test_file.unlink()
+        print(f"[TEST] Deleted test file: {test_file}")
+
+
+# Inicializar el TestClient DESPUÉS del fixture
+client = TestClient(app)
+
+
+def test_index_document_endpoint():
+    """
+    Test POST /index/{document_id} with a real file created by the fixture.
+    """
+    response = client.post("/index/test_doc_endpoint")
+    assert response.status_code == 202
+
+    data = response.json()
+    assert "correlationId" in data
+    assert "message" in data
+
+    print(f"[TEST] /index/test_doc_endpoint passed ✓ - correlationId: {data['correlationId']}")
+
+
+def test_index_stats_endpoint():
+    """
+    Verify that GET /index/stats returns correct indexing statistics.
+    """
+    response = client.get("/index/stats")
+    assert response.status_code == 200
+
+    data = response.json()
+
+    # Campos esperados del endpoint
+    assert data.get("status") == "ok"
+    assert "documentsIndexed" in data
+    assert "chunksStored" in data
+    assert "vectorDb" in data
+
+    print(
+        f"[TEST] /index/stats passed ✓ - docs: {data['documentsIndexed']}, "
+        f"chunks: {data['chunksStored']}"
+    )
