+++ tests/integration/indexing/test_index_local.py
